requirement.txt
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
joblib==1.4.2
numpy==2.1.3
PyJWT==2.9.0
scikit-learn==1.5.2
scipy==1.14.1
sqlparse==0.5.1
threadpoolctl==3.5.0
django-redis==5.4.0
redis==5.2.0
pytest==8.3.3
